% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfr.R
\name{CFR}
\alias{CFR}
\title{Helper function for calculating case-fatality rate}
\usage{
CFR(C, D)
}
\arguments{
\item{C}{Cases vector}

\item{D}{Deaths vector}
}
\value{
A case-fatality rate estimate and 95\% confidence limits
}
\description{
Helper function for calculating case-fatality rate
}
\examples{
\dontrun{
# used in get_cfr() function
}


}
\keyword{internal}
