% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_data_correction.R
\name{check_for_data_correction}
\alias{check_for_data_correction}
\title{Check for data corrections of X-times magnitude - data.table style}
\usage{
check_for_data_correction(
  dat,
  var = "daily_cases",
  magnitude = 10,
  min_count = 10
)
}
\arguments{
\item{dat}{data set}

\item{var}{variable for which to check for corrections. Default is \code{"daily_cases"}}

\item{magnitude}{magnitude of difference that qualifies as a data correction. Default is \code{10}.}

\item{min_count}{minimum count of var. Default is \code{10}.}
}
\value{
Data set with data correction observations removed
}
\description{
Check for data corrections of X-times magnitude - data.table style
}
\examples{
\dontrun{
check_for_data_correction(dat = get_nat_counts, var = "daily_cases", magnitude = 10)
}
}
