% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r0.R
\name{estR0_out}
\alias{estR0_out}
\title{Helper function for calculating R_0 - data.table style}
\usage{
estR0_out(x, incubation_days = 3)
}
\arguments{
\item{x}{Input dataset}

\item{incubation_days}{Number of days from infection to symptoms}
}
\value{
Pulls the time-series state-level testing data directly from covid19india.org.
}
\description{
Helper function for calculating R_0 - data.table style
}
\examples{
\dontrun{
estR0_out_dt(x = get_nat_counts())
}

}
\keyword{internal}
