% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_latest.R
\name{extract_latest}
\alias{extract_latest}
\title{Helper function}
\usage{
extract_latest(dat, group = place, clmns = c("total_tests", "tpr", "ppt"))
}
\arguments{
\item{dat}{data set}

\item{group}{place variable}

\item{clmns}{columns to be extracted}
}
\value{
Data set of recent observations of selected variables
}
\description{
Helper function
}
\examples{
\dontrun{
extract_latest(dat = get_all_data())
}
}
