% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Pull all covid19india count, test, and vaccine data for states and nation}
\usage{
get_all_data(keep_nat = TRUE)
}
\arguments{
\item{keep_nat}{Keep the national data as well. Default is \code{FALSE}}
}
\value{
Pulls the district-level time-series case, death, and recovered data directly from covid19india.org.
}
\description{
Pull all covid19india count, test, and vaccine data for states and nation
}
\examples{
\dontrun{
get_all_data()
}
}
