% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfr.R
\name{get_cfr}
\alias{get_cfr}
\title{Calculate case_fataility rate}
\usage{
get_cfr(dat)
}
\arguments{
\item{dat}{Input dataset. Expects \code{total_cases} and \code{total_deaths} variables}
}
\value{
Calculates a case-fatality rate estimate and corresponding 95\% confidence interval
}
\description{
Calculate case_fataility rate
}
\examples{
\dontrun{
# get_nat_counts() \%>\% get_cfr()
}

}
