% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics_tables.R
\name{get_metrics_tables}
\alias{get_metrics_tables}
\title{Create metrics tables}
\usage{
get_metrics_tables(top20 = NULL)
}
\arguments{
\item{top20}{Vector of state abbreviations for top 20 table}
}
\value{
Creates metrics tables for use in covind19.org
}
\description{
Create metrics tables
}
\examples{
\dontrun{
dat <- get_all_data()
cfr    <- get_cfr(dat) \%>\% distinct()
r0_est <- get_r_est(dat)
tabs   <- get_metrics_tables()

tabs$full
}
}
