% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nat_counts.R
\name{get_nat_counts}
\alias{get_nat_counts}
\title{Pull covid19india national time series data}
\usage{
get_nat_counts(
  path = "https://api.covid19india.org/csv/latest/case_time_series.csv",
  raw = FALSE,
  corr_check = FALSE,
  useDT = TRUE
)
}
\arguments{
\item{path}{The URL path for the data. Default: https://api.covid19india.org/csv/latest/case_time_series.csv}

\item{raw}{Pull raw unaltered data. Default is \code{FALSE}}

\item{corr_check}{Check for data correction. Default is \code{FALSE}}

\item{useDT}{Use data.table backend rather than tidyverse. Default is \code{TRUE}}
}
\value{
Pulls the time-series case, death, and recovered data directly from covid19india.org.
}
\description{
Pull covid19india national time series data
}
\examples{
\dontrun{
get_nat_counts()
}

}
