% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r0.R
\name{get_r0}
\alias{get_r0}
\title{Calculate r0}
\usage{
get_r0(
  dat,
  inc_var = daily_cases,
  daily_filter = 0,
  total_filter = 50,
  min_date = "2020-03-23"
)
}
\arguments{
\item{dat}{Input dataset. Expects \code{daily_cases}, \code{total_cases}, and \code{place} columns}

\item{daily_filter}{Theshold for minimum daily cases. Default = 0.}

\item{total_filter}{Threshold for minimum total cases reported to date. Default = 50.}

\item{min_date}{Threshold for earliest date to report R_0. Default = "2020-03-23".}
}
\value{
Pulls the time-series state-level testing data directly from covid19india.org.
}
\description{
Calculate r0
}
\examples{
\dontrun{
# used in get_r0() function
}

}
