% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_est.R
\name{get_r_est}
\alias{get_r_est}
\title{Helper function for pulling latest R estimates}
\usage{
get_r_est(dat)
}
\arguments{
\item{dat}{data set containing R estimates}
}
\value{
Pulls 7-day trailing average R estimates and 95\% confidence intervals
}
\description{
Helper function for pulling latest R estimates
}
\examples{
\dontrun{
get_all_data() \%>\% get_r_est()
}
}
