% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_counts.R
\name{get_state_counts}
\alias{get_state_counts}
\title{Pull covid19india state}
\usage{
get_state_counts(
  path = "https://api.covid19india.org/csv/latest/state_wise_daily.csv",
  raw = FALSE,
  keep_nat = FALSE,
  corr_check = TRUE
)
}
\arguments{
\item{path}{The URL path for the data. Default: \verb{https://api.covid19india.org/csv/latest/state_wise_daily.csv}}

\item{raw}{Pull raw unaltered data. Default is \code{FALSE}}

\item{keep_nat}{Keep the national data as well. Default is \code{FALSE}}

\item{corr_check}{Check for data correction. Default is \code{TRUE}}
}
\value{
Pulls the time-series case, death, and recovered data directly from covid19india.org.
}
\description{
Pull covid19india state
}
\examples{
\dontrun{
get_state_counts()
}
}
