% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_tests.R
\name{get_state_tests}
\alias{get_state_tests}
\title{Pull covid19india state-level testing data}
\usage{
get_state_tests(
  path = "https://api.covid19india.org/csv/latest/statewise_tested_numbers_data.csv",
  raw = FALSE
)
}
\arguments{
\item{path}{The URL path for the data. Default: https://api.covid19india.org/csv/latest/statewise_tested_numbers_data.csv}

\item{raw}{Pull raw unaltered data. Default is FALSE}
}
\value{
Pulls the time-series state-level testing data directly from covid19india.org.
}
\description{
Pull covid19india state-level testing data
}
\examples{
\dontrun{
get_state_tests()
}
}
