% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_vax.R
\name{get_state_vax}
\alias{get_state_vax}
\title{Pull covid19india state-level vaccine data}
\usage{
get_state_vax(
  path = "https://api.covid19india.org/csv/latest/vaccine_doses_statewise_v2.csv",
  raw = FALSE,
  keep_nat = TRUE
)
}
\arguments{
\item{path}{The URL path for the data. Default: https://api.covid19india.org/csv/latest/vaccine_doses_statewise_v2.csv}

\item{raw}{Pull raw unaltered data. Default is FALSE}

\item{keep_nat}{Keep national level data? Default is TRUE}
}
\value{
Pulls the time-series state-level vaccine data directly from covid19india.org.
}
\description{
Pull covid19india state-level vaccine data
}
\examples{
\dontrun{
get_state_counts()
}
}
