% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop}
\alias{pop}
\title{List of places, abbreviations, and populations}
\format{
A data frame with 3 variables: \code{place}, \code{abbrev},
\code{population}
\describe{
\item{place}{The name of the place}
\item{abbrev}{The abbreviations corresponding to place}
\item{population}{The population size}
}
}
\usage{
pop
}
\description{
List of places, abbreviations, and populations
}
\keyword{datasets}
