% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateWiseTimeSeries}
\alias{stateWiseTimeSeries}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) State wise time series Outbreak Dataset}
\format{
A data.frame object
}
\source{
Coronavirus India Database \href{https://www.covid19india.org/}{website}
}
\usage{
stateWiseTimeSeries
}
\description{
State wise time series outbreak dataset updated daily
}
\details{
The dataset contains the time series numbers of confirmed, recovered and death of the Coronavirus pandemic cases in India state wise. The data includes the following fields:

- Date - Date

- Status - Variable like (Confirmed, recovered and deaths)

- TT - Total of variables

- .... - Rest of the columns are states with coded as state key
}
\examples{

data(stateWiseTimeSeries)

head(stateWiseTimeSeries)

}
\keyword{COVID19}
\keyword{India}
\keyword{coronavirus}
\keyword{datasets}
