% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeSeriesIndia}
\alias{timeSeriesIndia}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) India time series Outbreak Dataset}
\format{
A data.frame object
}
\source{
Coronavirus India Database \href{https://www.covid19india.org/}{website}
}
\usage{
timeSeriesIndia
}
\description{
India time series outbreak dataset updated daily
}
\details{
The dataset contains the time series numbers of confirmed, recovered and death of the Coronavirus pandemic cases in India. The data includes the following fields:

- Date - Date

- Daily.Confirmed - Daily Confirmed cases

- Total.Confirmed - Cumulative confirmed cases

- Daily.Recovered - Daily Recovered cases

- Total.Recovered - Cumulative Recovered cases

- Daily.Deceased - Daily Deaths cases

- Total.Deceased- Cumulative Deaths cases
}
\examples{

data(timeSeriesIndia)

head(timeSeriesIndia)

}
\keyword{COVID19}
\keyword{India}
\keyword{coronavirus}
\keyword{datasets}
