% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAllData.R
\name{updateAllData}
\alias{updateAllData}
\title{Update all the datasets}
\usage{
updateAllData()
}
\description{
This function updates all the datasets in the repository in a single call
}
\author{
Shubhram Pandey
}
