% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateStateWise.R
\name{updateStateWise}
\alias{updateStateWise}
\title{Update state wise data}
\usage{
updateStateWise(branch = "master")
}
\arguments{
\item{branch}{default branch is master}
}
\description{
This function updates state wise data on daily basis from website
}
\author{
Shubhram Pandey
}
