% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTimeSeries.R
\name{updateTimeSeriesIndia}
\alias{updateTimeSeriesIndia}
\title{Update daily data in time series format}
\usage{
updateTimeSeriesIndia(branch = "master")
}
\arguments{
\item{branch}{default branch is master}
}
\description{
This function updates data in a time series format
}
\examples{
updateTimeSeriesIndia()
}
\author{
Shubhram Pandey
}
