% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{biblioteca}
\alias{biblioteca}
\title{Install and load libraries from CRAN and GitHub}
\usage{
biblioteca(libs)
}
\arguments{
\item{libs}{A character vector of libraries to install and load. Use \code{user_name/package_name} syntax for GitHub packages.}
}
\value{
Install and load packages from CRAN and GitHub.
}
\description{
Install and load libraries from CRAN and GitHub
}
\examples{
\dontrun{
biblioteca(c("tidyverse", "janitor", "maxsal/covid19india"))
}
}
