% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r0.R
\name{estR0_out}
\alias{estR0_out}
\title{Helper function for calculating R_0}
\usage{
estR0_out(dat)
}
\arguments{
\item{dat}{Input dataset}
}
\value{
Pulls the time-series state-level testing data directly from covid19india.org.
}
\description{
Helper function for calculating R_0
}
\examples{
\dontrun{
# used in get_r0() function
}

}
\keyword{internal}
