% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_latest.R
\name{extract_latest}
\alias{extract_latest}
\title{Helper function}
\usage{
extract_latest(
  data,
  group = place,
  cols = c("total_tests", "tpr", "dbl", "ppt")
)
}
\arguments{
\item{group}{place variable}

\item{cols}{columns to be extracted}

\item{dat}{data set}
}
\value{
Data set of recent observations of selected variables
}
\description{
Helper function
}
\examples{
\dontrun{

}
}
